moved event start filters and updated to new huerta filter
